refactor: move pages files to folders
