+++ src/app/router.tsx
@@ -1,9 +1,9 @@
 // src/app/router.tsx
 import { Routes, Route } from 'react-router-dom'
 
-import HomePage from '../pages/HomePage'
-import ChannelPage from '../pages/ChannelPage'
-import VideoPlayerPage from '../pages/VideoPlayerPage'
+import HomePage from '../pages/homePage/HomePage'
+import ChannelPage from '../pages/channelPage/ChannelPage'
+import VideoPlayerPage from '../pages/VideoPlayerPage/VideoPlayerPage'
 
 const AppRouter = () => {
   return (
