--- src/pages/HomePage.tsx
@@ -1,11 +1,0 @@
-import { Link } from 'react-router-dom'
-
-const HomePage = () => {
-  return (
-    <h2>
-      HomePage <Link to="/"></Link>
-    </h2>
-  )
-}
-
-export default HomePage
